requirements.txt
python-telegram-bot==20.3
PyPDF2==3.0.1
faiss-cpu
sentence-transformers
requests
python-dotenv
gigachat